GPUCellListSPH

Documentation for GPUCellListSPH.

GPUCellListSPH.GPUCellList

GPUCellListSPH.GPUCellList

GPUCellListSPH.SPHProblem

GPUCellListSPH.W_2d!

GPUCellListSPH.cellmap_2d!

GPUCellListSPH.cellmap_3d!

GPUCellListSPH.completed_∂v∂t!

GPUCellListSPH.completed_∂v∂t_3d!

GPUCellListSPH.cspmcorr!

GPUCellListSPH.dpcreg!

GPUCellListSPH.dpcreg!

GPUCellListSPH.dpcreg_3d!

GPUCellListSPH.fbmolforce!

GPUCellListSPH.fillcells_naive_2d!

GPUCellListSPH.fillcells_naive_3d!

GPUCellListSPH.makedf

GPUCellListSPH.makedf

GPUCellListSPH.neib_external_2d!

GPUCellListSPH.neib_external_3d!

GPUCellListSPH.neib_internal_2d!

GPUCellListSPH.neib_internal_3d!

GPUCellListSPH.neighborlist

GPUCellListSPH.neighborlist

GPUCellListSPH.neighborlistview

GPUCellListSPH.partialupdate!

GPUCellListSPH.partialupdate!

GPUCellListSPH.pressure!

GPUCellListSPH.stepsolve!

GPUCellListSPH.timesolve!

GPUCellListSPH.update!

GPUCellListSPH.update!

GPUCellListSPH.update_all!

GPUCellListSPH.update_all_3d!

GPUCellListSPH.update_dpcreg!

GPUCellListSPH.update_dpcreg_3d!

GPUCellListSPH.update_vp∂v∂tΔt!

GPUCellListSPH.update_vp∂v∂tΔt_3d!

GPUCellListSPH.update_xpvΔt!

GPUCellListSPH.update_xpvΔt_3d!

GPUCellListSPH.update_Δt!

GPUCellListSPH.update_ρ!

GPUCellListSPH.update_ρ_3d!

GPUCellListSPH.writecsv

GPUCellListSPH.writecsv

GPUCellListSPH.xsphcorr!

GPUCellListSPH.Δt_stepping

GPUCellListSPH.Δt_stepping_3d

GPUCellListSPH.мaxpairs_2d

GPUCellListSPH.мaxpairs_3d

GPUCellListSPH.∂v∂t!

GPUCellListSPH.∂v∂t!

GPUCellListSPH.∂v∂t_3d!

GPUCellListSPH.∂v∂tpF!

GPUCellListSPH.∂v∂tpF!

GPUCellListSPH.∂v∂tpF_3d!

GPUCellListSPH.∂Π∂t!

GPUCellListSPH.∂Π∂t!

GPUCellListSPH.∂Π∂t_3d!

GPUCellListSPH.∂ρ∂tDDT!

GPUCellListSPH.∂ρ∂tDDT!

GPUCellListSPH.∂ρ∂tDDT_3d!

GPUCellListSPH.∇W_2d!

GPUCellListSPH.∑W_2d!

GPUCellListSPH.∑W_2d!

GPUCellListSPH.∑W_3d!

GPUCellListSPH.∑W∑∇W_2d!

GPUCellListSPH.∑∇W_2d!

GPUCellListSPH.∑∇W_2d!

GPUCellListSPH.∑∇W_2d!

GPUCellListSPH.∑∇W_3d!

Docs

GPUCellListSPH.GPUCellList — Type

GPUCellList(points, cellsize, dist; mppcell = 0, mpairs = 0)

Make cell list structure.

GPUCellListSPH.update! — Function

update!(c::GPUCellList)

Full update cell grid.

GPUCellListSPH.partialupdate! — Function

partialupdate!(c::GPUCellList, fillzero::Bool = true)

Update only distance

GPUCellListSPH.neighborlist — Function

neighborlist(c::GPUCellList)

List of pairs with distance.

GPUCellListSPH.∑W_2d! — Function

∑W_2d!(∑W, pairs, sphkernel, H⁻¹)

Compute sum of kernel values for each particles pair in list. Add to ∑W. See SPHKernels.jl for details.

GPUCellListSPH.∑∇W_2d! — Function

∑∇W_2d!(∑∇W, ∇W, pairs, points, kernel, H⁻¹)

Compute gradients. Add sum to ∑∇W and update ∇W. See SPHKernels.jl for details.

∑∇W_2d!(∑∇W, pairs, points, kernel, H⁻¹)

Compute gradients. Add sum to ∑∇W. See SPHKernels.jl for details.

GPUCellListSPH.∂ρ∂tDDT! — Function

∂ρ∂tDDT!(∑∂ρ∂t,  ∇W, pairs, points, h, m₀, δᵩ, c₀, γ, g, ρ₀, ρ, v, ptype)

Compute ∂ρ∂t - density derivative includind density diffusion.

\[ \frac{\partial \rho_i}{\partial t} = \sum m_j \textbf{v}_{ij} \cdot \nabla_i W_{ij} + \delta_{\Phi} h c_0 \sum \Psi_{ij} \cdot \nabla_i W_{ij} \frac{m_j}{\rho_j} \\ \Psi_{ij} = 2 (\rho_{ij}^T - \rho_{ij}^H) \frac{\textbf{r}_{ij}}{r_{ij}^2 + \eta^2} \\ \rho_{ij}^H = \rho_0 \left( \sqrt[\gamma]{\frac{P_{ij}^H + 1}{C_b}} - 1\right) \\ P_{ij}^H = \rho_0 g z_{ij} \]

$z_{ij}$ - vertical distance.

GPUCellListSPH.∂Π∂t! — Function

∂Π∂t!(∑∂Π∂t, ∇W, pairs, points, h, ρ, α, v, c₀, m₀)

Compute ∂Π∂t - artificial viscosity. Add to ∑∂Π∂t

\[ \Pi_{ij} = \begin{cases} \frac{- \alpha \overline{c}_{ij} \mu_{ij} + \beta \mu_{ij}^2 }{\overline{\rho}_{ij}} & \textbf{v}_{ij}\cdot \textbf{r}_{ij} < 0 \\ 0 & otherwise \end{cases} \\ \\ \mu_{ij} = \frac{h \textbf{v}_{ij}\cdot \textbf{r}_{ij}}{r_{ij}^2 + \eta^2} \\ \\ \overline{c}_{ij} = \frac{c_i + c_j}{2} \\ \\ \overline{\rho}_{ij} = \frac{\rho_i + \rho_j}{2} \\ \\ \beta = 0 \c_{ij} = c_0 \m_i = m_j = m_0 \]

Artificial viscosity part of momentum equation.

\[ \frac{\partial \textbf{v}_i}{\partial t} = - \sum m_j \Pi_{ij} \nabla_i W_{ij}\]

J. Monaghan, Smoothed Particle Hydrodynamics, “Annual Review of Astronomy and Astrophysics”, 30 (1992), pp. 543-574.

GPUCellListSPH.∂v∂t! — Function

∂v∂t!(∑∂v∂t,  ∇Wₙ, pairs, m, ρ, c₀, γ, ρ₀)

The momentum equation (without dissipation).

\frac{\partial \textbf{v}i}{\partial t} = - \sum mj \left( \frac{pi}{\rho^2i} + \frac{pj}{\rho^2j} \right) \nablai W{ij}

Missing docstring.

Missing docstring for GPUCellListSPH.pressure. Check Documenter's build log for details.

GPUCellListSPH.∂v∂tpF! — Function

∂v∂tpF!(∑∂v∂t, pairs, points, s, H)

Add surface tension to ∑∂v∂t. Modified.

A. Tartakovsky and P. Meakin, Phys. Rev. E 72 (2005)

GPUCellListSPH.dpcreg! — Function

dpcreg!(∑Δvdpc, v, ρ, P, pairs, points, sphkernel, l₀, Pmin, Pmax, Δt, λ, dpckernlim)

Dynamic Particle Collision (DPC) correction.

\[\delta \textbf{v}_i^{DPC} = \sum k_{ij}\frac{m_j}{m_i + m_j}v_{ij}^{coll} + \frac{\Delta t}{\rho_i}\sum \phi_{ij} \frac{2V_j}{V_i + V_j}\frac{p_{ij}^b}{r_{ij}^2 + \eta^2}\textbf{r}_{ij} \\ (v_{ij}^{coll} , \quad \phi_{ij}) = \begin{cases} (\frac{\textbf{v}_{ij}\cdot \textbf{r}_{ij}}{r_{ij}^2 + \eta^2}\textbf{r}_{ji}, \quad 0) & \textbf{v}_{ij}\cdot \textbf{r}_{ij} < 0 \\ (0, \quad 1) & otherwise \end{cases} \\ p_{ij}^b = \tilde{p}_{ij} \chi_{ij} \\ \tilde{p}_{ij} = max(min(\lambda |p_i + p_j|, \lambda p_{max}), p_{min}) \\ \chi_{ij} = \sqrt{\frac{\omega({r}_{ij}, l_0)}{\omega(l_0/2, l_0)}} \\ k_{ij} = \begin{cases} \chi_{ij} & 0.5 \le {r}_{ij}/l_0 < 1 \\ 1 & {r}_{ij}/l_0 < 0.5 \end{cases} \]

Mojtaba Jandaghian, Herman Musumari Siaben, Ahmad Shakibaeinia, Stability and accuracy of the weakly compressible SPH with particle regularization techniques https://arxiv.org/pdf/2110.10076.pdf

GPUCellListSPH.makedf — Function

makedf(prob::SPHProblem; vtkvars = ["Density", "Acceleration", "Velocity"])

Make DataFrame from SPH Problem.

GPUCellListSPH.writecsv — Function

writecsv(prob::SPHProblem, path; vtkvars = ["Density", "Acceleration", "Velocity"])

write CSV file.