GPUCellListSPH

Documentation for GPUCellListSPH.

GPUCellListSPH.kernel_cellmap_2d!

GPUCellListSPH.kernel_cellpnum_2d!

GPUCellListSPH.kernel_cellthreadmap_2d!

GPUCellListSPH.kernel_fillcells_cspwn_2d!

GPUCellListSPH.kernel_fillcells_naive_2d!

GPUCellListSPH.kernel_neib_external_2d!

GPUCellListSPH.kernel_neib_internal_2d!

GPUCellListSPH.partialupdate!

GPUCellListSPH.update!

GPUCellListSPH.kernel_cellmap_2d! — Method

Map each point to cell.

GPUCellListSPH.kernel_cellpnum_2d! — Method

Number of points in each cell.

GPUCellListSPH.kernel_cellthreadmap_2d! — Method

Fill cell list with cell. Each thread find starting point in sorted array and iterate only in region for first dimension of cell grid.

GPUCellListSPH.kernel_fillcells_cspwn_2d! — Method

Fill cell list with cell. Each cell iterate over pcell vector.

GPUCellListSPH.kernel_fillcells_naive_2d! — Method

Fill cell list with cell. Naive approach.

GPUCellListSPH.kernel_neib_external_2d! — Method

Find all pairs with another cell shifted on offset.

GPUCellListSPH.kernel_neib_internal_2d! — Method

Find all pairs with distance < h

GPUCellListSPH.partialupdate! — Method

partialupdate!(c::GPUCellList)

Update only distance

GPUCellListSPH.update! — Method

update!(c::GPUCellList)

Full update cell grid.