GPUCellListSPH

Documentation for GPUCellListSPH.

GPUCellListSPH.GPUCellList

GPUCellListSPH.GPUCellList

GPUCellListSPH.cellmap_2d!

GPUCellListSPH.completed_∂v∂t!

GPUCellListSPH.fillcells_naive_2d!

GPUCellListSPH.neib_external_2d!

GPUCellListSPH.neib_internal_2d!

GPUCellListSPH.neighborlist

GPUCellListSPH.neighborlist

GPUCellListSPH.partialupdate!

GPUCellListSPH.partialupdate!

GPUCellListSPH.pressure

GPUCellListSPH.pressure

GPUCellListSPH.stepsolve!

GPUCellListSPH.timesolve!

GPUCellListSPH.update!

GPUCellListSPH.update!

GPUCellListSPH.update_all!

GPUCellListSPH.update_vp∂v∂tΔt!

GPUCellListSPH.update_xpvΔt!

GPUCellListSPH.update_Δt!

GPUCellListSPH.update_ρ!

GPUCellListSPH.Δt_stepping

GPUCellListSPH.мaxpairs_2d

GPUCellListSPH.∂v∂t!

GPUCellListSPH.∂v∂t!

GPUCellListSPH.∂v∂tpF!

GPUCellListSPH.∂Π∂t!

GPUCellListSPH.∂Π∂t!

GPUCellListSPH.∂ρ∂tDDT!

GPUCellListSPH.∂ρ∂tDDT!

GPUCellListSPH.∑W_2d!

GPUCellListSPH.∑W_2d!

GPUCellListSPH.∑W∑∇W_2d!

GPUCellListSPH.∑∇W_2d!

GPUCellListSPH.∑∇W_2d!

GPUCellListSPH.GPUCellList — Method

GPUCellList(points, cellsize, dist; mppcell = 0, mpairs = 0)

Make cell list structure.

GPUCellListSPH.cellmap_2d! — Method

cellmap_2d!(pcell, cellpnum, points,  h, offset)

Map each point to cell and count number of points in each cell.

For each coordinates cell number calculated:

csᵢ = size(cellpnum, 1) 
p₁  =  (x₁ - offset₁) * h₁⁻¹
pᵢ₁ = ceil(min(max(p₁, 1), csᵢ))

GPUCellListSPH.completed_∂v∂t! — Method

completed_∂vᵢ∂t!(∑∂v∂t, ∑∂Π∂t,  gvec, gfac)

Add gravity and artificial viscosity to the momentum equation.

GPUCellListSPH.fillcells_naive_2d! — Method

fillcells_naive_2d!(celllist, cellpnum, pcell)

Fill cell list with cell. Naive approach. No bound check. Values in pcell list shoid be in range of cellpnum and celllist.

GPUCellListSPH.neib_external_2d! — Method

neib_external_2d!(pairs, cnt, cellpnum, points, celllist, offset, dist)

Find all pairs with another cell shifted on offset.

GPUCellListSPH.neib_internal_2d! — Method

neib_internal_2d!(pairs, cnt, cellpnum, points, celllist, dist)

Find all pairs with distance < h in one cell.

GPUCellListSPH.neighborlist — Method

neighborlist(c::GPUCellList)

List of pairs with distance.

GPUCellListSPH.partialupdate! — Method

partialupdate!(c::GPUCellList)

Update only distance

GPUCellListSPH.pressure — Method

pressure(ρ, c₀, γ, ρ₀)

Equation of State in Weakly-Compressible SPH

GPUCellListSPH.stepsolve! — Function

stepsolve!(prob::SPHProblem, n::Int = 1; timecall = nothing, timestepping = false, timelims = (-Inf, +Inf))

Make n itarations.

timestepping - call Δt_stepping for adjust Δt

timelims - minimal and maximum values for Δt

GPUCellListSPH.timesolve! — Method

timesolve!(prob::SPHProblem; batch = 10, timeframe = 1.0, writetime = 0, path = nothing, pvc = false, timestepping = false, timelims = (-Inf, +Inf), anim = false)

Make simulation by batch iterations within timeframe.

writetime - time interval for write vtk / animation.

path - path to export directory.

anim - make animation.

showframe - show animation each frame.

GPUCellListSPH.update! — Method

update!(c::GPUCellList)

Full update cell grid.

GPUCellListSPH.update_all! — Method

update_all!(ρ, ρΔt½, v, vΔt½, x, xΔt½, ∑∂ρ∂t, ∑∂v∂t,  Δt, ρ₀, isboundary, ml)

GPUCellListSPH.update_vp∂v∂tΔt! — Method

update_vp∂v∂tΔt!(v, ∑∂v∂t, Δt, ml)

GPUCellListSPH.update_xpvΔt! — Method

update_xpvΔt!(x, v, Δt, ml)

GPUCellListSPH.update_Δt! — Method

update_ρ!(ρ, ∑∂ρ∂t, Δt, ρ₀, isboundary)

Update ρ, v, x at timestep Δt and derivatives: ∑∂ρ∂t, ∑∂v∂t.

GPUCellListSPH.update_ρ! — Method

update_ρ!(ρ, ∑∂ρ∂t, Δt, ρ₀, isboundary)

GPUCellListSPH.Δt_stepping — Method

Δt_stepping(buf, a, v, points, c₀, h, CFL, timelims)

GPUCellListSPH.мaxpairs_2d — Method

мaxpairs_2d(cellpnum)

Maximum number of pairs.

GPUCellListSPH.∂v∂t! — Method

∂v∂t!(∑∂v∂t,  ∇Wₙ, pairs, m, ρ, c₀, γ, ρ₀)

The momentum equation (without dissipation).

GPUCellListSPH.∂v∂tpF! — Method

∂v∂tpF!(∑∂v∂t, pairs, points, s, H)

Add surface tension to ∑∂v∂t. Modified.

A. Tartakovsky and P. Meakin, Phys. Rev. E 72 (2005)

GPUCellListSPH.∂Π∂t! — Method

∂Π∂t!(∑∂Π∂t, ∇Wₙ, pairs, points, h, ρ, α, v, c₀, m₀)

Compute ∂Π∂t - artificial viscosity.

GPUCellListSPH.∂ρ∂tDDT! — Method

∂ρ∂tDDT!(∑∂ρ∂t,  ∇Wₙ, pairs, points, h, m₀, δᵩ, c₀, γ, g, ρ₀, ρ, v, MotionLimiter)

Compute ∂ρ∂t - density derivative includind density diffusion.

GPUCellListSPH.∑W_2d! — Method

∑W_2d!(sumW, pairs, sphkernel, H⁻¹)

Compute ∑W for each particles pair in list.

GPUCellListSPH.∑W∑∇W_2d! — Method

∑W∑∇W_2d!(∑W, ∑∇W, ∇Wₙ, pairs, points, sphkernel, H⁻¹)

Compute ∑W ans gradient for each particles pair in list.

GPUCellListSPH.∑∇W_2d! — Method

∑∇W_2d!(sum∇W, ∇Wₙ, pairs, points, kernel, H⁻¹)

Compute gradients.

API

GPUCellListSPH.GPUCellList — Type

GPUCellList(points, cellsize, dist; mppcell = 0, mpairs = 0)

Make cell list structure.

GPUCellListSPH.update! — Function

update!(c::GPUCellList)

Full update cell grid.

GPUCellListSPH.partialupdate! — Function

partialupdate!(c::GPUCellList)

Update only distance

GPUCellListSPH.neighborlist — Function

neighborlist(c::GPUCellList)

List of pairs with distance.

GPUCellListSPH.∑W_2d! — Function

∑W_2d!(sumW, pairs, sphkernel, H⁻¹)

Compute ∑W for each particles pair in list.

GPUCellListSPH.∑∇W_2d! — Function

∑∇W_2d!(sum∇W, ∇Wₙ, pairs, points, kernel, H⁻¹)

Compute gradients.

GPUCellListSPH.∂ρ∂tDDT! — Function

∂ρ∂tDDT!(∑∂ρ∂t,  ∇Wₙ, pairs, points, h, m₀, δᵩ, c₀, γ, g, ρ₀, ρ, v, MotionLimiter)

Compute ∂ρ∂t - density derivative includind density diffusion.

GPUCellListSPH.∂Π∂t! — Function

∂Π∂t!(∑∂Π∂t, ∇Wₙ, pairs, points, h, ρ, α, v, c₀, m₀)

Compute ∂Π∂t - artificial viscosity.

GPUCellListSPH.∂v∂t! — Function

∂v∂t!(∑∂v∂t,  ∇Wₙ, pairs, m, ρ, c₀, γ, ρ₀)

The momentum equation (without dissipation).

GPUCellListSPH.pressure — Function

pressure(ρ, c₀, γ, ρ₀)

Equation of State in Weakly-Compressible SPH