Autodocs API

GPUCellListSPH.GPUCellList — Method

GPUCellList(points, cellsize, dist; mppcell = 0, mpairs = 0)

Make cell list structure.

GPUCellListSPH.SPHProblem — Type

SPHProblem(system::GPUCellList, h::Float64, H::Float64, sphkernel::AbstractSPHKernel, ρ, v, ptype, ρ₀::Float64, m₀::Float64, Δt::Float64, α::Float64, g::Float64, c₀::Float64, γ, δᵩ::Float64, CFL::Float64; s::Float64 = 0.0)

SPH simulation data structure.

GPUCellListSPH.W_2d! — Method

W_2d!(W, pairs, sphkernel, H⁻¹)

Compute kernel values for each particles pair in list. Update W. See SPHKernels.jl for details.

GPUCellListSPH.cellmap_2d! — Method

cellmap_2d!(pcell, cellpnum, points,  h, offset)

Map each point to cell and count number of points in each cell.

For each coordinates cell number calculated:

csᵢ = size(cellpnum, 1) 
p₁  =  (x₁ - offset₁) * h₁⁻¹
pᵢ₁ = ceil(min(max(p₁, 1), csᵢ))

GPUCellListSPH.cellmap_3d! — Method

cellmap_3d!(pcell, cellpnum, points,  h, offset)

Map each point to cell and count number of points in each cell.

For each coordinates cell number calculated:

GPUCellListSPH.completed_∂v∂t_3d! — Method

completed_∂vᵢ∂t!(∑∂v∂t, ∑∂Π∂t,  gvec, gfac)

Add gravity and artificial viscosity to the momentum equation.

GPUCellListSPH.cspmcorr! — Method

cspmcorr!(∑ρcspm1, ∑ρcspm2, ρ, m₀, pairs, points, sphkernel, H⁻¹)

Corrected Smoothed Particle Method (CSPM) Density Renormalisation.

\[ \rho_{i}^{norm} = \frac{\sum m_j W}{\sum \frac{m_j}{\rho_j} W}\]

Chen JK, Beraun JE, Carney TC (1999) A corrective smoothed particle method for boundary value problems in heat conduction. Int. J. Num. Meth. Engng. https://doi.org/10.1002/(SICI)1097-0207(19990920)46:2<231::AID-NME672>3.0.CO;2-K

GPUCellListSPH.dpcreg! — Method

dpcreg!(∑Δvdpc, v, ρ, P, pairs, points, sphkernel, l₀, Pmin, Pmax, Δt, λ, dpckernlim)

Dynamic Particle Collision (DPC) correction. Replace all values and update ∑Δvdpc.

\[\delta \textbf{v}_i^{DPC} = \sum k_{ij}\frac{m_j}{m_i + m_j}v_{ij}^{coll} + \frac{\Delta t}{\rho_i}\sum \phi_{ij} \frac{2V_j}{V_i + V_j}\frac{p_{ij}^b}{r_{ij}^2 + \eta^2}\textbf{r}_{ij} \\ (v_{ij}^{coll} , \quad \phi_{ij}) = \begin{cases} (\frac{\textbf{v}_{ij}\cdot \textbf{r}_{ij}}{r_{ij}^2 + \eta^2}\textbf{r}_{ji}, \quad 0) & \textbf{v}_{ij}\cdot \textbf{r}_{ij} < 0 \\ (0, \quad 1) & otherwise \end{cases} \\ p_{ij}^b = \tilde{p}_{ij} \chi_{ij} \\ \tilde{p}_{ij} = max(min(\lambda |p_i + p_j|, \lambda p_{max}), p_{min}) \\ \chi_{ij} = \sqrt{\frac{\omega({r}_{ij}, l_0)}{\omega(l_0/2, l_0)}} \\ k_{ij} = \begin{cases} \chi_{ij} & 0.5 \le {r}_{ij}/l_0 < 1 \\ 1 & {r}_{ij}/l_0 < 0.5 \end{cases} \]

Mojtaba Jandaghian, Herman Musumari Siaben, Ahmad Shakibaeinia, Stability and accuracy of the weakly compressible SPH with particle regularization techniques https://arxiv.org/pdf/2110.10076.pdf

GPUCellListSPH.dpcreg_3d! — Method

dpcreg!(∑Δvdpc, v, ρ, P, pairs, points, sphkernel, l₀, Pmin, Pmax, Δt, λ, dpckernlim)

Dynamic Particle Collision (DPC) correction.

Mojtaba Jandaghian, Herman Musumari Siaben, Ahmad Shakibaeinia, Stability and accuracy of the weakly compressible SPH with particle regularization techniques https://arxiv.org/pdf/2110.10076.pdf

GPUCellListSPH.fbmolforce! — Method

fbmolforce!(∑∂v∂t, pairs, points, d, r₀, ptype)

The repulsive force exerted by the virtual particle on the fluid particle.

\[F = D * \frac{\left( (\frac{r_0}{\textbf{r}_{ij}})^{n_1} - (\frac{r_0}{\textbf{r}_{ij}})^{n_2}\right)}{r_{ij}^2}\]

Rapaport, 2004

n₁ = 12

n₂ = 4

GPUCellListSPH.fillcells_naive_2d! — Method

fillcells_naive_2d!(celllist, cellpnum, pcell)

Fill cell list with cell. Naive approach. No bound check. Values in pcell list shoid be in range of cellpnum and celllist.

GPUCellListSPH.fillcells_naive_3d! — Method

fillcells_naive_3d!(celllist, cellpnum, pcell)

Fill cell list with cell. Naive approach. No bound check. Values in pcell list shoid be in range of cellpnum and celllist.

GPUCellListSPH.makedf — Method

makedf(prob::SPHProblem; vtkvars = ["Density", "Acceleration", "Velocity"])

Make DataFrame from SPH Problem.

GPUCellListSPH.neib_external_2d! — Method

neib_external_2d!(pairs, cnt, cellpnum, points, celllist, offset, dist)

Find all pairs with another cell shifted on offset.

GPUCellListSPH.neib_external_3d! — Method

neib_external_3d!(pairs, cnt, cellpnum, points, celllist, offset, dist)

Find all pairs with another cell shifted on offset.

GPUCellListSPH.neib_internal_2d! — Method

neib_internal_2d!(pairs, cnt, cellpnum, points, celllist, dist)

Find all pairs with distance < h in one cell.

GPUCellListSPH.neib_internal_3d! — Method

neib_internal_2d!(pairs, cnt, cellpnum, points, celllist, dist)

Find all pairs with distance < h in one cell.

GPUCellListSPH.neighborlist — Method

neighborlist(c::GPUCellList)

List of pairs with distance.

GPUCellListSPH.neighborlistview — Method

neighborlistview(c::GPUCellList)

List of pairs with distance.

GPUCellListSPH.partialupdate! — Function

partialupdate!(c::GPUCellList, fillzero::Bool = true)

Update only distance

GPUCellListSPH.pressure! — Method

pressure!(P, ρ, ρ₀, c₀, γ)

Equation of State in Weakly-Compressible SPH.

\[P = c_0^2 \rho_0 * \left[ \left( \frac{\rho}{\rho_0} \right)^{\gamma} \right]\]

GPUCellListSPH.stepsolve! — Function

stepsolve!(prob::SPHProblem, n::Int = 1; timecall = nothing, timestepping = false, timelims = (-Inf, +Inf))

Make n itarations.

timestepping - call Δt_stepping for adjust Δt

timelims - minimal and maximum values for Δt

GPUCellListSPH.symplectic_update! — Method

symplectic_update!(ρ, ρΔt½, v, vΔt½, x, xΔt½, ∑∂ρ∂t, ∑∂v∂t,  Δt, ρ₀, isboundary, ml)

Symplectic Position Verlet scheme.

Parshikov et al, 2000

Leimkuhler and Matthews, 2016

GPUCellListSPH.timesolve! — Method

timesolve!(prob::SPHProblem; batch = 10, 
timeframe = 1.0, 
writetime = 0, 
path = nothing, 
pvc = false, 
vtkvars = ["Acceleration", "Velocity", "Pressure"],
timestepping = false, 
timelims = (-Inf, +Inf), 
anim = false,
plotsettings = Dict(:leg => false))

Make simulation by batch iterations within timeframe.

writetime - time interval for write vtk / animation.

path - path to export directory.

pvc - make PVD file.

vtkvars - variables for export, full list: ["Acceleration", "Velocity", "Pressure", "Density", "∑W", "∑∇W", "DPC"]

anim - make animation.

showframe - show animation each frame.

plotsettings - keywords for plotting.

GPUCellListSPH.update! — Function

update!(c::GPUCellList)

Full update cell grid.

GPUCellListSPH.update_all_3d! — Method

update_all!(ρ, ρΔt½, v, vΔt½, x, xΔt½, ∑∂ρ∂t, ∑∂v∂t,  Δt, ρ₀, isboundary, ml)

GPUCellListSPH.update_dpcreg! — Method

update_dpcreg!(v, x, ∑Δvdpc, Δt, ptype)

Update velocity and position.

GPUCellListSPH.update_dpcreg_3d! — Method

update_dpcreg!(v, x, ∑Δvdpc, Δt, isboundary)

Update velocity and position.

GPUCellListSPH.update_vp∂v∂tΔt! — Method

update_vp∂v∂tΔt!(v, ∑∂v∂t, Δt, ptype)

Update vlocity.

\[\textbf{v} = \textbf{v} + \frac{\partial \textbf{v}}{\partial t} * \Delta t\]

GPUCellListSPH.update_vp∂v∂tΔt_3d! — Method

update_vp∂v∂tΔt!(v, ∑∂v∂t, Δt, ml)

GPUCellListSPH.update_xpvΔt! — Method

update_xpvΔt!(x, v, Δt, ml)

\[\textbf{r} = \textbf{r} + \textbf{v} * \Delta t\]

GPUCellListSPH.update_xpvΔt_3d! — Method

update_xpvΔt!(x, v, Δt, ml)

GPUCellListSPH.update_xsphcorr! — Method

update_xsphcorr!(v, ∑Δvxsph, ptype)

Update velocity.

GPUCellListSPH.update_ρ_3d! — Method

update_ρ!(ρ, ∑∂ρ∂t, Δt, ρ₀, isboundary)

GPUCellListSPH.update_ρp∂ρ∂tΔt! — Method

update_ρp∂ρ∂tΔt!(ρ, ∑∂ρ∂t, Δt, ρ₀, ptype)

Update dencity.

\[\rho = \rho + \frac{\partial \rho}{\partial t} * \Delta t\]

GPUCellListSPH.writecsv — Method

writecsv(prob::SPHProblem, path; vtkvars = ["Density", "Acceleration", "Velocity"])

write CSV file.

GPUCellListSPH.xsphcorr! — Method

xsphcorr!(∑Δvxsph, v, ρ, W, pairs, m₀)

The XSPH correction.

\[ \hat{\textbf{v}_{i}} = - \epsilon \sum m_j \frac{\textbf{v}_{ij}}{\overline{\rho}_{ij}} W_{ij} \]

Monaghan JJ (1989) On the problem of penetration in particle methods. J Comput Phys. https://doi.org/10.1016/0021-9991(89)90032-6

Carlos Alberto Dutra Fraga Filho, Reflective Boundary Conditions Coupled With the SPH Method for the Three-Dimensional Simulation of Fluid-Structure Interaction With Solid Boundaries

GPUCellListSPH.Δt_stepping — Method

Δt_stepping(buf, a, v, points, c₀, h, CFL, timelims)

GPUCellListSPH.Δt_stepping_3d — Method

Δt_stepping(buf, a, v, points, c₀, h, CFL, timelims)

GPUCellListSPH.мaxpairs_2d — Method

мaxpairs_2d(cellpnum)

Maximum number of pairs.

GPUCellListSPH.мaxpairs_3d — Method

мaxpairs_3d(cellpnum)

Maximum number of pairs.

GPUCellListSPH.∂v∂t! — Method

∂v∂t!(∑∂v∂t,  ∇W, pairs, m₀, ρ, c₀, γ, ρ₀)

The momentum equation (without dissipation and gravity). Add to ∑∂v∂t.

\frac{\partial \textbf{v}i}{\partial t} = - \sum mj \left( \frac{pi}{\rho^2i} + \frac{pj}{\rho^2j} \right) \nablai W{ij}

GPUCellListSPH.∂v∂t_3d! — Method

∂v∂t!(∑∂v∂t,  ∇Wₙ, pairs, m, ρ, c₀, γ, ρ₀)

The momentum equation (without dissipation).

GPUCellListSPH.∂v∂t_addgrav! — Method

∂v∂t_addgrav!(∑∂v∂t, gvec)

Add gravity to the momentum equation.

GPUCellListSPH.∂v∂t_av! — Method

∂v∂t_av!(∑∂v∂t, ∇W, pairs, points, h, ρ, α, v, c₀, m₀)

Compute artificial viscosity part of ∂v∂t. Add to ∑∂v∂t.

\[ \Pi_{ij} = \begin{cases} \frac{- \alpha \overline{c}_{ij} \mu_{ij} + \beta \mu_{ij}^2 }{\overline{\rho}_{ij}} & \textbf{v}_{ij}\cdot \textbf{r}_{ij} < 0 \\ 0 & otherwise \end{cases} \\ \\ \mu_{ij} = \frac{h \textbf{v}_{ij}\cdot \textbf{r}_{ij}}{r_{ij}^2 + \eta^2} \\ \\ \overline{c}_{ij} = \frac{c_i + c_j}{2} \\ \\ \overline{\rho}_{ij} = \frac{\rho_i + \rho_j}{2} \\ \\ \beta = 0 \c_{ij} = c_0 \m_i = m_j = m_0 \]

Artificial viscosity part of momentum equation.

\[ \frac{\partial \textbf{v}_i}{\partial t} = - \sum m_j \Pi_{ij} \nabla_i W_{ij}\]

J. Monaghan, Smoothed Particle Hydrodynamics, “Annual Review of Astronomy and Astrophysics”, 30 (1992), pp. 543-574.

GPUCellListSPH.∂v∂t_visc! — Method

∂v∂t_visc!(∑∂v∂t,  ∇W, pairs, m, ρ, c₀, γ, ρ₀)

Compute laminar shear stresse part of ∂v∂t. Add to ∑∂v∂t.

\[\frac{\partial \textbf{v}_i}{\partial t} = \sum \frac{m_j}{\rho_j} \left( 2 \nu_i \frac{\textbf{r}_{ij} \cdot \nabla_i W_{ij} }{r_{ij}^2} \right) \textbf{v}_{ij}\]

GPUCellListSPH.∂v∂tpF! — Method

∂v∂tpF!(∑∂v∂t, pairs, points, s, H)

Add surface tension to ∑∂v∂t. Modified.

A. Tartakovsky and P. Meakin, Phys. Rev. E 72 (2005)

GPUCellListSPH.∂v∂tpF_3d! — Method

∂v∂tpF!(∑∂v∂t, pairs, points, s, H)

Add surface tension to ∑∂v∂t. Modified.

A. Tartakovsky and P. Meakin, Phys. Rev. E 72 (2005)

GPUCellListSPH.∂Π∂t_3d! — Method

∂Π∂t!(∑∂Π∂t, ∇Wₙ, pairs, points, h, ρ, α, v, c₀, m₀)

Compute ∂Π∂t - artificial viscosity.

GPUCellListSPH.∂ρ∂tDDT! — Method

∂ρ∂tDDT!(∑∂ρ∂t,  ∇W, pairs, points, h, m₀, δᵩ, c₀, γ, g, ρ₀, ρ, v, ptype)

Compute ∂ρ∂t - density derivative includind density diffusion. Replace all values and update ∑∂ρ∂t.

\[ \frac{\partial \rho_i}{\partial t} = \sum m_j \textbf{v}_{ij} \cdot \nabla_i W_{ij} + \delta_{\Phi} h c_0 \sum \Psi_{ij} \cdot \nabla_i W_{ij} \frac{m_j}{\rho_j} \\ \Psi_{ij} = 2 (\rho_{ij}^T - \rho_{ij}^H) \frac{\textbf{r}_{ij}}{r_{ij}^2 + \eta^2} \\ \rho_{ij}^H = \rho_0 \left( \sqrt[\gamma]{\frac{P_{ij}^H + 1}{C_b}} - 1\right) \\ P_{ij}^H = \rho_0 g z_{ij} \]

$z_{ij}$ - vertical distance.

GPUCellListSPH.∂ρ∂tDDT_3d! — Method

∂ρ∂tDDT!(∑∂ρ∂t,  ∇Wₙ, pairs, points, h, m₀, δᵩ, c₀, γ, g, ρ₀, ρ, v, MotionLimiter)

Compute ∂ρ∂t - density derivative includind density diffusion.

GPUCellListSPH.∇W_2d! — Method

∇W_2d!(∇W, pairs, points, kernel, H⁻¹)

Compute gradients. Update ∇W. See SPHKernels.jl for details.

GPUCellListSPH.∑W_2d! — Method

∑W_2d!(∑W, pairs, sphkernel, H⁻¹)

Compute sum of kernel values for each particles pair in list. Add to ∑W. See SPHKernels.jl for details.

GPUCellListSPH.∑W_3d! — Method

∑W_2d!(sumW, pairs, sphkernel, H⁻¹)

Compute ∑W for each particles pair in list.

GPUCellListSPH.∑∇W_2d! — Method

∑∇W_2d!(∑∇W, pairs, points, kernel, H⁻¹)

Compute gradients. Add sum to ∑∇W. See SPHKernels.jl for details.

GPUCellListSPH.∑∇W_2d! — Method

∑∇W_2d!(∑∇W, ∇W, pairs, points, kernel, H⁻¹)

Compute gradients. Add sum to ∑∇W and update ∇W. See SPHKernels.jl for details.

GPUCellListSPH.∑∇W_3d! — Method

∑∇W_3d!(∑∇W, ∇W, pairs, points, kernel, H⁻¹)

Compute gradients.